app/oauth.py
from authlib.integrations.flask_client import OAuth
import os

oauth = OAuth()  # Create the OAuth object without initializing

def init_oauth(app):
    oauth.init_app(app)
    oauth.register(
    name="google",
        client_id=os.getenv("GOOGLE_CLIENT_ID"),
        client_secret=os.getenv("GOOGLE_CLIENT_SECRET"),
        authorize_url="https://accounts.google.com/o/oauth2/auth",
        access_token_url="https://oauth2.googleapis.com/token",
        client_kwargs={"scope": "openid email profile"},
        api_base_url="https://www.googleapis.com/oauth2/v1/",
        userinfo_endpoint="https://openidconnect.googleapis.com/v1/userinfo",
        jwks_uri="https://www.googleapis.com/oauth2/v3/certs"  # Add the jwks_uri
    )