requirements.txt
Flask==2.2.3
requests==2.28.2
flask_oauthlib==0.9.5
supabase==0.3.1